:art: feat: remove serversideprops from edit page
